[Rename] DiaryCalendar 디렉토리 경로 수정 - calendar -> diary-calendar - App.tsx 주석 삭제
