+++ src/App.tsx
@@ -9,7 +9,7 @@ import DiaryDetail from './pages/diary-detail';
 import Profile from './pages/profile';
 import PublicDiary from './pages/public-diary';
 import WriteHereMaP from './pages/write-here-map';
-import DiaryCalendar from './pages/calendar';
+import DiaryCalendar from './pages/diary-calendar';
 import GoodBye from './pages/good-bye';
 
 function App() {
@@ -23,7 +23,6 @@ function App() {
           <Route path="/diary" element={<Diary />} />
           <Route path="/landing" element={<Landing />} />
           <Route path="/diary-detail" element={<DiaryDetail />} />
-          {/* <Route path="/diary-register" element={<DiaryRegister />} /> */}
           <Route path="/profile" element={<Profile />} />
           <Route path="/public-diary" element={<PublicDiary />} />
           <Route path="/write-here-map" element={<WriteHereMaP />} />
